Send emails when the dewar state changes
Users need to be informed when their dewars arrive, depart, are filled
with LN2 or go missing. This change adds a command line program that
subscribes to the pucktracker websocket and sends the emails when it
detects a relevant action. Printing of the dewar slip when dewars arrive
is also handled by this program.